refactor: move project into src
