+++ README.md
@@ -47,56 +47,14 @@ The project is in early stages. To run the frontend locally:
 ## Project Structure
 
 ```
-├── core/                      # App-wide singleton services and utilities
-│   ├── services/
-│   │   ├── auth.service.ts
-│   │   ├── sermon-ai.service.ts
-│   │   ├── scripture.service.ts
-│   │   ├── storage.service.ts
-│   │   └── voice.service.ts
-│   ├── guards/
-│   │   └── auth.guard.ts
-│   └── interceptors/
-│       └── http.interceptor.ts
-
-├── shared/                    # Shared UI components, pipes, and directives
-│   ├── components/
-│   │   ├── sermon-card/
-│   │   └── editor-toolbar/
-│   ├── pipes/
-│   │   └── scripture-highlight.pipe.ts
-│   └── directives/
-│       └── autofocus.directive.ts
-
-├── features/                  # App features grouped by domain
-│   ├── home/
-│   │   └── home.page.ts
-│   ├── auth/
-│   │   ├── login.page.ts
-│   │   ├── register.page.ts
-│   │   └── profile.page.ts
-│   ├── sermon/
-│   │   ├── create/
-│   │   │   └── create-sermon.page.ts
-│   │   ├── editor/
-│   │   │   └── sermon-editor.page.ts
-│   │   ├── list/
-│   │   │   └── sermon-list.page.ts
-│   │   └── detail/
-│   │       └── sermon-detail.page.ts
-│   ├── community/
-│   │   ├── community-feed.page.ts
-│   │   └── sermon-share.page.ts
-│   └── reflection/
-│       ├── prayer.page.ts
-│       └── reflection.page.ts
-
-├── assets/
-│   └── i18n/
-├── environments/
-│   ├── environment.ts
-│   └── environment.prod.ts
-├── app.config.ts
-├── app.routes.ts
-└── main.ts
+src/
+  ├── app/
+  │   ├── core/               # Singleton services and utilities
+  │   ├── shared/             # Shared UI components
+  │   └── features/           # App features grouped by domain
+  ├── environments/
+  │   ├── environment.ts
+  │   └── environment.prod.ts
+  ├── index.html
+  └── main.ts
 ```
