+++ angular.json
@@ -7,14 +7,14 @@
     "sermon-companion": {
       "projectType": "application",
       "root": "",
-      "sourceRoot": "",
+      "sourceRoot": "src",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-angular:browser",
           "options": {
             "outputPath": "dist/sermon-companion",
-            "index": "index.html",
-            "main": "main.ts",
+            "index": "src/index.html",
+            "main": "src/main.ts",
             "tsConfig": "tsconfig.json",
             "assets": [],
             "styles": [],
@@ -24,8 +24,8 @@
             "production": {
               "fileReplacements": [
                 {
-                  "replace": "environments/environment.ts",
-                  "with": "environments/environment.prod.ts"
+                  "replace": "src/environments/environment.ts",
+                  "with": "src/environments/environment.prod.ts"
                 }
               ]
             }
