+++ src/app/core/services/scripture.service.ts
