+++ src/app/core/services/sermon-ai.service.ts
