+++ src/app/core/services/voice.service.ts
