+++ src/app/features/auth/register.page.css
