+++ src/app/shared/components/editor-toolbar/index.ts
