+++ src/app/shared/components/sermon-card/index.ts
