+++ src/app/shared/pipes/scripture-highlight.pipe.ts
